Silence a bunch of MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@23033 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/nitf/nitfdes.c
@@ -277,7 +277,7 @@ retry:
     }
     else
     {
-        char* pachData = (char*)VSIMalloc(psSegInfo->nSegmentSize);
+        char* pachData = (char*)VSIMalloc((size_t)psSegInfo->nSegmentSize);
         if (pachData == NULL )
         {
             CPLDebug("NITF", "Cannot allocate " CPL_FRMT_GUIB " bytes DES data",
@@ -285,7 +285,7 @@ retry:
         }
         else if( VSIFSeekL( psFile->fp, psSegInfo->nSegmentStart,
                     SEEK_SET ) != 0
-            || VSIFReadL( pachData, 1, psSegInfo->nSegmentSize,
+            || VSIFReadL( pachData, 1, (size_t)psSegInfo->nSegmentSize,
                         psFile->fp ) != psSegInfo->nSegmentSize )
         {
             CPLDebug("NITF",
